Refactor project structure and update namespaces
Significant refactoring to improve code organization:
- Updated namespaces from `TravelExpertMVC` to `TravelExpertData`
- Created new `TravelExpertData` project and updated references
- Moved EF Core package references to `TravelExpertData`
- Renamed `PackagesManager` to `PackagesRepository`
- Enabled nullable reference types and implicit usings
- Updated solution file to include `TravelExpertData`
- Modified `Error.cshtml` to use `TravelExpertMVC.Models.ErrorViewModel`
- Removed `TravelExpertMVC.Models` import from `_ViewImports.cshtml`
- Updated `PackagesController` to use `PackagesRepository`
- Updated migration files for new namespaces and structure
- Updated `TravelExpertContext` and model classes
- Updated project files to reflect new structure and dependencies
