+++ TravelExpertData/Models/Fee.cs
@@ -1,10 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using Microsoft.EntityFrameworkCore;
 
-namespace TravelExpertMVC.Models;
+namespace TravelExpertData.Models;
 
 public partial class Fee
 {
