+++ TravelExpertData/Models/User.cs
@@ -1,10 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using Microsoft.EntityFrameworkCore;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using Microsoft.EntityFrameworkCore;
 
-namespace TravelExpertMVC.Models;
+namespace TravelExpertData.Models;
 
 [Keyless]
 public partial class User
