+++ TravelExpertMVC/Views/Shared/Error.cshtml
@@ -1,4 +1,4 @@
-﻿@model ErrorViewModel
+﻿@model TravelExpertMVC.Models.ErrorViewModel
 @{
     ViewData["Title"] = "Error";
 }
