+++ TravelExpertMVC/Views/_ViewImports.cshtml
@@ -1,3 +1,2 @@
 ﻿@using TravelExpertMVC
-@using TravelExpertMVC.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
